TST: move test to test_base_indexer
